common: separate memory allocators
Make the memory allocators more visible so that
users can easier replace them with a custom file.
