+++ CONTRIBUTORS
@@ -48,3 +48,4 @@ SoonGeon Noh <nors.nsg@gmail.com>
 Giseong Ji <jiggyjiggy0323@gmail.com>
 KunYoung Park <rjsdud3263@gmail.com>
 Wang SiMiao 王思淼 <wangsimiao1@xiaomi.com>
+Jackson Hu <huming2207@gmail.com>
