+++ src/common/tvgCommon.h
@@ -90,32 +90,9 @@ namespace tvg {
     #define TVG_DELETE(PAINT) \
     if (PAINT->refCnt() == 0) delete(PAINT)
 
-    //custom memory allocators
-    template<typename T = void*>
-    static inline T malloc(size_t size)
-    {
-        return static_cast<T>(std::malloc(size));
-    }
-
-    template<typename T = void*>
-    static inline T calloc(size_t nmem, size_t size)
-    {
-        return static_cast<T>(std::calloc(nmem, size));
-    }
-
-    template<typename T = void*>
-    static inline T realloc(void* ptr, size_t size)
-    {
-        return static_cast<T>(std::realloc(ptr, size));
-    }
-
-    template<typename T = void*>
-    static inline void free(void* ptr)
-    {
-        std::free(ptr);
-    }
-
     extern int engineInit;
 }
+
+#include "tvgAllocator.h"
 
 #endif //_TVG_COMMON_H_
